% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsmooth.R
\name{fsmooth}
\alias{fsmooth}
\title{fsmooth Function to convert a vector of data points into a continuous functional object}
\usage{
fsmooth(
  dt,
  M = 6,
  genLfun.fd = 4,
  centered = FALSE,
  nbasis = NULL,
  breaks = NULL,
  lambda = NULL,
  rangeval = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{dt}{object of class data.table with dimensions number of data point per number of samples}

\item{M}{spline order (norder)}

\item{genLfun.fd}{Lfdobj for fdPar}

\item{centered}{center by r before smoothing}

\item{nbasis}{allow user to set number of basis functions}

\item{breaks}{provide a coarser breaks vector}

\item{lambda}{smoothing penalty; if NULL a default is chosen}

\item{rangeval}{override range(r) if desired}

\item{verbose}{verbose true or false}

\item{...}{other arguments for smooth.basis function}
}
\value{
A list with the fd objects and the original matrix and r vector
}
\description{
fsmooth Function to convert a vector of data points into a continuous functional object
}
