% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpaFun-package.R
\docType{package}
\name{SpaFun-package}
\alias{SpaFun}
\alias{SpaFun-package}
\title{SpaFun: Spatial Point Pattern Analysis with FDA}
\description{
Function to analyse TCGA dataset.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/revalescente/SpaFun}
  \item \url{https://revalescente.github.io/SpaFun/}
  \item Report bugs at \url{https://github.com/revalescente/SpaFun/issues}
}

}
\author{
\strong{Maintainer}: Valerio Reffo \email{valerio.reffo@phd.unipd.it} (\href{https://orcid.org/0009-0006-8690-7018}{ORCID})

Authors:
\itemize{
  \item Valerio Reffo \email{valerio.reffo@phd.unipd.it} (\href{https://orcid.org/0009-0006-8690-7018}{ORCID})
}

}
\keyword{internal}
